function_app/requirements.txt
azure-functions>=1.17.0,<2.0.0
azure-identity>=1.15.0,<2.0.0
azure-keyvault-secrets>=4.7.0,<5.0.0
azure-storage-blob>=12.19.0,<13.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
